refactor: Update Zabbix agent cloud-config to install Docker
Update the cloud-config file for Zabbix agent to install Docker
using the official Docker repository instead of the default
Debian package. This ensures that the latest version of Docker
is installed. Also add comments to explain the purpose of the
file and the variables that will be replaced by Terraform.